refactored the package and changed private variables to protected
