+++ src/main/java/org/bcit/comp2522/project/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.bcit.comp2522.project;
 
 public class Main {
     public static void main(String[] args) {
