+++ src/main/java/org/bcit/comp2522/project/Player.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.bcit.comp2522.project;
 
 public class Player {
