+++ src/main/java/org/bcit/comp2522/project/Sprite.java
@@ -1,11 +1,11 @@
-package org.example;
+package org.bcit.comp2522.project;
 
 
 public abstract class Sprite implements Drawable {
 
-  private final float xpos;
-  private final float ypos;
-  private final Window window;
+  protected final float xpos;
+  protected final float ypos;
+  protected final Window window;
 
   public Sprite(float xpos, float ypos, Window window) {
     this.xpos = xpos;
