+++ src/main/java/org/bcit/comp2522/project/Tower_Manager.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.bcit.comp2522.project;
 
 import java.util.ArrayList;
